requirements.txt
numpy
pandas
matplotlib
seaborn
sklearn
xgboost
catboost
nltk
wordcloud
requests
re
pathlib
pyyaml